apps: move VolumePlacementOpts definition to its own file
