+++ apps/glusterfs/volume_placement_opts.go
@@ -1,0 +1,45 @@
+//
+// Copyright (c) 2018 The heketi Authors
+//
+// This file is licensed to you under your choice of the GNU Lesser
+// General Public License, version 3 or any later version (LGPLv3 or
+// later), or the GNU General Public License, version 2 (GPLv2), in all
+// cases as published by the Free Software Foundation.
+//
+
+package glusterfs
+
+type VolumePlacementOpts struct {
+	v            *VolumeEntry
+	brickSize    uint64
+	numBrickSets int
+}
+
+func NewVolumePlacementOpts(v *VolumeEntry,
+	brickSize uint64, numBrickSets int) *VolumePlacementOpts {
+	return &VolumePlacementOpts{v, brickSize, numBrickSets}
+}
+
+func (vp *VolumePlacementOpts) BrickSizes() (uint64, float64) {
+	return vp.brickSize, float64(vp.v.Info.Snapshot.Factor)
+}
+
+func (vp *VolumePlacementOpts) BrickOwner() string {
+	return vp.v.Info.Id
+}
+
+func (vp *VolumePlacementOpts) BrickGid() int64 {
+	return vp.v.Info.Gid
+}
+
+func (vp *VolumePlacementOpts) SetSize() int {
+	return vp.v.Durability.BricksInSet()
+}
+
+func (vp *VolumePlacementOpts) SetCount() int {
+	return vp.numBrickSets
+}
+
+func (vp *VolumePlacementOpts) AverageFileSize() uint64 {
+	return vp.v.GetAverageFileSize()
+}
